Revert "Reapply "BPF: Enable libbpf build flag in trunk staging""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0554ccd5-d840-4bdc-a2ae-0755e49b4d3f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iadc8c8c3ab4914074317e1119f410275c46b5f86
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com