Revert "Add build flag to control if we turn on mainline aconfig..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05fb2bb8-dd38-4d97-a53a-1b1bda6e6d9b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2b84e1d61451e8ce5f113aa051dfecd9b1244117
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com