Revert "Enable dex container (DEX v41) for trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0e29929b-6576-438f-b8ea-55356251cede). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieef073a19ef5a9fc4b6f3ec134ee9fb93c39161e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com