Revert "Enable dex container (DEX v41) for trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/44b08bad-cb98-468a-b4d0-adc307ce45cf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0f346d90d8b89a7f6e8c63c5fc60695960d94e22
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com