Revert "Enable dex container (DEX v41) for trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/201dd683-e2b6-4c8c-aa9d-513b6976a970). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I99ec6af6403d02e9c5d90cb001735e979c89d1ad
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com