Revert "Enable dex container (DEX v41) for trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a9c568c8-8353-4d22-a203-058442c7f0d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4e51ac19b744276a7ff25b68467b7dea48f7d6ba
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com