Revert "Enable dex container (DEX v41) for trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/71865a68-7346-4329-a0bd-16aa53cf33d7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia710a34c85694537e5aa500aa0493f5ec1cebef4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com