Revert "Enable dex container (DEX v41) for trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/674d0bbd-096a-4077-a10c-57cc966774bd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie538804421b1f02fe4df14575036c89ba80fea5d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com