Revert "Rename the filename for native_power_notifications feature"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c17d4773-02fc-40d6-8828-1cef08e87675). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 386431128
Change-Id: Id535c1f04bf24ca2b5c106a3f313a86e76e33a0a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com