Revert "Rename the filename for native_power_notifications feature"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d50cff4a-c0bf-405b-a41a-a57963195657). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 386430899
Change-Id: I87589329c589327bb9084b7d097f745bde36770d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com