Revert "Move dc back button flags to staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/28123abd-a422-4614-8da8-d3dd77a8b305). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3214fdeaf5de1b887f47693d0bd2fc1bd34126b1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com