Revert "Enable dex container (DEX v41) for trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/366fb5a6-198a-40b5-950c-18c3ec36d182). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I92c1f9444a72579277630ee121d3ba0eb394759b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com