Revert "Enable dex container (DEX v41) for trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/58578be5-f7a6-41c8-8cfb-cdb8af410edf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic51d9d13dff7e706ef21eb549fa6091a13320117
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com