Revert "Enable RELEASE_R8_STORE_STORE_FENCE_CONSTRUCTOR_INLINING..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e59e2771-874c-4be7-9a19-99b94ebbbbd5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I588f1a3203af00e1003ad7fabc95d1302ed235da
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com