Revert "Enable RELEASE_R8_STORE_STORE_FENCE_CONSTRUCTOR_INLINING..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2ce788d4-4dc3-4a69-b7d3-81a032d43a9d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia4094e02fc62d9cc7f3e7bbb68244a7c09e7c669
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com