Revert "Enable RELEASE_R8_STORE_STORE_FENCE_CONSTRUCTOR_INLINING..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8eba40af-a5aa-4756-b1ba-6af8cb170ac7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic58cb09b57cfe7f65a1327e801a76eaad5698318
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com