Revert "Enable dex container (DEX v41) for trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/59b49b40-dd46-4170-9a3c-a26ed4583306). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I724fb1a16a7ddb99254f68af63e825b248128a42
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com