Revert "Enable RELEASE_R8_STORE_STORE_FENCE_CONSTRUCTOR_INLINING..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b41620d8-d934-431b-913c-6c37267db759). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib7fef08e5519a7dfb8611eca9d83fbe641b649af
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com