Revert "Revert "Enable RELEASE_R8_STORE_STORE_FENCE_CONSTRUCTOR_..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a76ac2a4-ff19-4456-904c-b08c80263552). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib8d989266e3d82629b57227c4a7bff3770f3a660
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com