Revert "Revert "Enable RELEASE_R8_STORE_STORE_FENCE_CONSTRUCTOR_..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a67246da-3440-4d54-9f0d-dae0a07d9766). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie61fe104cdb3014bce35852773e94b3ecff33095
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com