Revert "Revert "Enable RELEASE_R8_STORE_STORE_FENCE_CONSTRUCTOR_..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/882278a6-bff0-468c-9dc5-d118fee78454). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib44092fe4fd690961a263aae5b16b607f3bdd1d3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com