Revert "Add RELEASE_R8_ONLY_RUNTIME_VISIBLE_ANNOTATIONS flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6c2c04d1-1b65-40ca-8236-11e7991cbd65). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I182f0bc70d2ffaf8e6c2368df243926be625f5cb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com