Revert "Add RELEASE_R8_ONLY_RUNTIME_VISIBLE_ANNOTATIONS flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/03f99fea-6ee6-48ba-a94d-ce21b4b183bf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibb1c30e3e312ce6ee12b61fd403af9cbdee6602b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com