Revert "Delete sharded jarjar build flag in trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb34cbac-31de-4722-bdd8-495b737144f6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iec4a877abb6d739d31b0739e2504632bed33706f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com