Revert "Drop aosp_current alias."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/73de8649-3bec-4bd9-9acd-8964eece5fc7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4982fcc1721d03d230dc8484f07d7666e472044a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com