Revert "Drop aosp_current alias."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f2f72f6b-a901-4b10-b7f1-ddb55c82a41e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If518ea8f79e458252d0e488ea1910f7f7576a70f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com