Revert "Drop aosp_current alias."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/64462312-d3de-472e-8aa1-a26dd7b390f0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7ab0f4d72aa090bfbd8b395fe63cca3f7a0ae79c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com