Revert "Drop aosp_current alias."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1565c247-6f87-4510-b686-902be8852759). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie0dcaddb84cfe79d903e508c5547f94a2ba7ca8c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com