Revert "Launch soong-only cuttlefish builds"
This reverts commit 6bf195a2f7eb82320dab9ab7decab9ab76eb10d1.

Reason for revert: b/396761822 This change has broken at least two targets. The error message is very unclear, and it took full day of multiple engineer in our team a full day to identify this CL is the cause. I believe many more users are impacted. Can we implement this change in a less disruptive way, or perform a staged rollout for the opted-in module? Note that even Droidmonitor seems to have not been informed of this change, which means the user couldn't even get support from Droidmonitor chatroom.

Change-Id: I1941f7e156406f66d6d1f8b9a18384c87f30a599