Revert "Add RELEASE_R8_GLOBAL_CHECK_NOT_NULL_FLAGS build flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4cc3dfcd-34cb-4142-b5fe-13fc22c188e6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4777d0a35cd619f361414bca29f7dab18be9394c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com