Revert "Advance trunk-staging to sqlite 3.44.4"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/73647c25-a40d-471b-b085-6c88e95c8dd7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5f1b4d7df55dd0c0440dcc05efb31d930d8cf888
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com