Revert "Advance trunk-staging to sqlite 3.44.4"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1002a85d-0d45-4a59-8eb5-e1076ec5e8c9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I072437653cfac831d2bb7a96bf13500be4df000f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com