Revert "Advance trunk-staging to sqlite 3.44.4"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/49965195-7557-4ac1-a5d7-16d8c7e58899). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idd2ddce0011689ecc957826546156053e39bc73f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com