Revert "Advance trunk-staging to sqlite 3.44.4"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dc17db55-424d-481e-a693-c1015ac85a91). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2d57d36d7eee641a98d8a6528455cea613fd5f45
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com