Revert "Advance trunk-staging to sqlite 3.44.4"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/09d446bd-aa9e-4d40-b294-0dbafbfc7718). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6eb515846c9842b26bff657347adb1dc497d4898
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com