release: Adding flag native_power_notification to bp1a
ICarPowerServer_default_V1_1 VTS test failing because
the ICarPowerServer AIDL service is not registered.
This is due to the native_power_notifications flag being disabled.

This change enables the native_power_notifications flag in the bp1a build config, which allows notifications to be sent to native listeners for power state changes such as battery level, charging status, and screen on/off events.

Bug: 419066560
Test: vts_treble_vintf_framework_test

Change-Id: I3d1173e253f065afadc837ed5342db78eee9a57e
Signed-off-by: Akshay Mhaske <asantosh@amd.com>